Rename crypto.packaged -> crypto.signing
